requirements.txt
beautifulsoup4==4.11.1
bs4==0.0.1
pandas==1.4.3
pyodbc==4.0.34
python-dotenv==0.20.0
requests==2.28.1
Scrapy==2.6.2
selenium==4.3.0
wget==3.2
SQLAlchemy~=1.4.40
numba~=0.56.2
pymssql~=2.2.5